Restore timezone labels with offsets
